Streamline Docker deployment and improve build tools
Refactored build tools and modified Docker deployment process. Docker-compose-template was dropped in favor of a Go-based template builder and shell script. Also, Docker build task in Makefile has been updated to utilize the new deployment_builder script. The name of the config file was changed from example.yaml to deployment.yaml, with corresponding updates made in code references.

These alterations were made to simplify and enhance deployment procedures and allows for customization of the docker-compose setup. The independent script replaces the make-based builder, facilitating decoupling of build and deployment processes and making it easier to adapt to different environments and CI/CD workflows.