Add function to log and store image details
A new function 'WriteStringToFile' was added and the conditional statement to push Docker image was modified in the 'image_builder.go'. This change now allows for the image details (name and tag) to be written to a file and logged when the image is pushed.

This enhancement improves the logging and debugging practices by maintaining a clear record of image details. This will also help trace