Refactor code to follow Go naming conventions
All variable names in the 'template_builder.go' file have been updated to follow Golang's naming conventions. Before this change, constants were used for struct members and function parameters which is not recommended in Go. Now, the struct members and parameters are using CamelCase which is more idiomatic to Go. This refactor improves readability and adherence to good coding standards in Go.