Update project structure and submodule configuration
The changes refactored the project structure, adjusting the Go naming conventions in 'template_builder.go' to be more idiomatic. This includes changing constants to CamelCase in string formatting for improved readability and code style adherence. Also, the .gitmodule configuration was updated to use HTTPS instead of SSH due to problems with firewall blocking SSH URLs in some systems. The project's 'docker-compose.yml' file was altered to change the build volume path, reflecting changes in the project structure. Lastly, a line was added in 'Makefile' for automatic submodule updates to ensure that the CI submodule is always the latest version when building the docker image.